Build spoonacular microservice
